# docs/swagger.yaml
basePath: /api/v1
definitions:
  bill.Cart:
    properties:
      billDetails:
        items:
          $ref: '#/definitions/bill.CartDetails'
        type: array
      billId:
        type: string
      customerId:
        type: integer
      employee:
        type: integer
      entryDate:
        type: string
      finishDate:
        type: string
      totalBill:
        type: integer
    type: object
  bill.CartDetails:
    properties:
      billId:
        type: string
      price:
        type: integer
      productId:
        type: integer
      qty:
        type: integer
      subTotal:
        type: integer
    type: object
  entity.BillDetails:
    properties:
      billId:
        type: string
      id:
        type: integer
      product:
        type: integer
      productPrice:
        type: integer
      qty:
        type: integer
    type: object
  entity.Customer:
    properties:
      address:
        type: string
      id:
        type: integer
      name:
        type: string
      phoneNumber:
        type: string
    type: object
  entity.Employee:
    properties:
      address:
        type: string
      department:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      password:
        type: string
      phoneNumber:
        type: string
    type: object
...
  entity.Transaction:
    properties:
      billDate:
        type: string
      billDetails:
        items:
          $ref: '#/definitions/entity.BillDetails'
        type: array
      customer:
        $ref: '#/definitions/entity.Customer'
      employee:
        $ref: '#/definitions/entity.Employee'
      entryDate:
        type: string
      finishDate:
        type: string
      id:
        type: integer
      totalBill:
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    email: sdesain25@gmail.com
    name: API Support
    url: https://uqonstd.xyz/
  description: A documentation how to access the api's routes on your laundry application
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Laundry API Documentation
  version: "1.0"
paths:
...
  /transactions:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all transactions
      parameters:
      - description: Filter by start date
        in: query
        name: startDate
        type: string
      - description: Filter by end date
        in: query
        name: endDate
        type: string
      - description: Filter by product name
        in: query
        name: productName
        type: string
      - default: Bearer <Add access token here>
        description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Transaction data
          schema:
            $ref: '#/definitions/entity.Transaction'
        "400":
          description: Bad Request
      security:
      - ApiKeyAuth: []
      summary: List all transactions
      tags:
      - Transactions
    post:
      consumes:
      - application/json
      description: Create a new bill with the provided cart details
      parameters:
      - description: Cart data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/bill.Cart'
      - default: Bearer <Add access token here>
        description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Cart data
          schema:
            $ref: '#/definitions/bill.Cart'
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
      security:
      - ApiKeyAuth: []
      summary: Create a new bill
      tags:
      - Transactions
  /transactions/{id_bill}:
    get:
      consumes:
      - application/json
      description: Retrieve a specific bill by its ID
      parameters:
      - description: Bill ID
        in: path
        name: id_bill
        required: true
        type: string
      - default: Bearer <Add access token here>
        description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Bill data
          schema:
            $ref: '#/definitions/entity.Transaction'
        "400":
          description: Bad Request
        "404":
          description: Not Found
      security:
      - ApiKeyAuth: []
      summary: Get bill by ID
      tags:
      - Transactions
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"